requirements.txt
feedparser==6.0.11
requests==2.32.3
python-dateutil==2.9.0.post0
openai==1.40.6
